iio: backend: add get regmap
Considering the need of a particular ADC/DAC IP implementation,
as i.e. the need to operate with the synchronous 4 bit parallel
bus implemented in ad3552r, allowing AXI regmap to be used from the
final fornt-end reduces growing of the backend code with custom
features needed only for a particular front-end driver.

Signed-off-by: Angelo Dureghello <adureghello@baylibre.com>